Add preliminary support for MacOS X in builder

+++ unix/opt/build-opt-unix.sh
@@ -22,6 +22,8 @@ if [ $GEM_SET_DEBUG ]; then
 fi
 set -e
 
+#Everyone has at least two cores
+NPROC=2
 OQ_ROOT=/tmp/build-openquake-dist
 OQ_REL=qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqq
 OQ_PREFIX=${OQ_ROOT}/${OQ_REL}/openquake
@@ -46,6 +48,12 @@ elif [ "$BUILD_OS" == "redhat" ]; then
     sudo yum -y upgrade
     sudo yum -y groupinstall 'Development Tools'
     sudo yum -y install autoconf libtool sqlite-devel readline-devel zlib-devel bzip2-devel wget xz git which
+elif [ "$BUILD_OS" == "macosx" ]; then
+    command -v xcode-select &> /dev/null || {
+        echo -e "!! Please install $i first." >&2
+        exit 1
+    }
+    sudo xcode-select --install || true
 else
     echo "Build OS not uspported"
     exit 1
@@ -58,18 +66,25 @@ cp install.sh build/
 
 cd build/src
 
-wget -nc https://www.openssl.org/source/openssl-1.0.2h.tar.gz
-wget -nc https://www.python.org/ftp/python/2.7.11/Python-2.7.11.tar.xz
-wget -nc http://www.hdfgroup.org/ftp/HDF5/current/src/hdf5-1.8.17.tar.gz
-wget -nc https://github.com/libgeos/libgeos/archive/3.5.0.tar.gz
-wget -nc https://bootstrap.pypa.io/get-pip.py
+curl -LOz openssl-1.0.2h.tar.gz https://www.openssl.org/source/openssl-1.0.2h.tar.gz
+curl -LOz Python-2.7.11.tar.xz https://www.python.org/ftp/python/2.7.11/Python-2.7.11.tar.xz
+curl -LOz hdf5-1.8.17.tar.gz http://www.hdfgroup.org/ftp/HDF5/current/src/hdf5-1.8.17.tar.gz
+curl -LOz get-pip.py https://bootstrap.pypa.io/get-pip.py
+
+if [ "$BUILD_OS" != "macosx" ]; then
+    curl -LOz 3.5.0.tar.gz https://github.com/libgeos/libgeos/archive/3.5.0.tar.gz
+fi
 
 cd ..
 
 cat <<EOF >> $OQ_PREFIX/requirements.txt
+pkgconfig==1.1.0
+Cython==0.23.4
 futures==3.0.5
 mock==1.3.0
-h5py==2.6.0
+# h5py must be installed after everything elese
+# to avoid SSL errors on MacOS X
+# h5py==2.6.0
 nose==1.3.7
 numpy==1.11.0
 pbr==1.8.0
@@ -90,15 +105,25 @@ export CPATH=\${PREFIX}/include
 export PATH=\${PREFIX}/bin:$PATH
 export OQ_SITE_CFG_PATH=\${PREFIX}/etc/openquake.cfg
 EOF
+if [ "$BUILD_OS" == "macosx" ]; then
+    cat <<EOF >> $OQ_PREFIX/env.sh
+export LC_ALL=en_US.UTF-8
+export LANG=en_US.UTF-8
+EOF
+fi
 
 source $OQ_PREFIX/env.sh
 
 if $CLEANUP; then rm -Rf openssl-1.0.2h; fi
 tar xvf src/openssl-1.0.2h.tar.gz
 cd openssl-1.0.2h/
-./config --prefix=$OQ_PREFIX shared
-make depend
-make
+if [ "$BUILD_OS" == "macosx" ]; then
+    ./Configure darwin64-x86_64-cc shared enable-ec_nistp_64_gcc_128 no-ssl2 no-ssl3 no-comp --prefix=$OQ_PREFIX
+else
+    ./config shared --prefix=$OQ_PREFIX
+fi
+make -j $NPROC depend
+make -j $NPROC
 make install
 cd ..
 
@@ -106,7 +131,7 @@ if $CLEANUP; then rm -Rf Python-2.7.11; fi
 tar xvf src/Python-2.7.11.tar.xz
 cd Python-2.7.11
 ./configure --prefix=$OQ_PREFIX
-make
+make -j $NPROC
 make install
 cd ..
 
@@ -115,30 +140,35 @@ tar xvf src/hdf5-1.8.17.tar.gz
 cd hdf5-1.8.17
 export HDF5_DIR=$OQ_PREFIX
 ./configure --prefix=$OQ_PREFIX
-make
+make -j $NPROC
 make install
 cd ..
 
-if $CLEANUP; then rm -Rf libgeos-3.5.0; fi
-tar xvf src/3.5.0.tar.gz
-cd libgeos-3.5.0
-# Workaround for an autogen.sh bug
-./autogen.sh || true
-./autogen.sh
-./configure --prefix=$OQ_PREFIX
-make
-make install
-cd ..
+if [ "$BUILD_OS" != "macosx" ]; then
+    if $CLEANUP; then rm -Rf libgeos-3.5.0; fi
+    tar xvf src/3.5.0.tar.gz
+    cd libgeos-3.5.0
+    # Workaround for an autogen.sh bug
+    ./autogen.sh || true
+    ./autogen.sh
+    ./configure --prefix=$OQ_PREFIX
+    make -j $NPROC
+    make install
+    cd ..
+fi
 
 python src/get-pip.py
 python $(which pip) install -r $OQ_PREFIX/requirements.txt
+# h5py must be installed after everything elese
+# to avoid SSL errors on MacOS X
+python $(which pip) install h5py==2.6.0
 
 for g in hazardlib engine;
 do 
     rm -Rf oq-${g}
     git clone --depth=1 -b $OQ_BRANCH https://github.com/gem/oq-${g}.git
     cd oq-${g}
-    declare OQ_${g^^}_DEV=$(git rev-parse --short HEAD)
+    declare OQ_$(echo $g | tr '[:lower:]' '[:upper:]')_DEV=$(git rev-parse --short HEAD)
     python setup.py install
     cd ..
 done
@@ -149,7 +179,13 @@ cp oq-engine/openquake.cfg $OQ_PREFIX/etc
 cp -R oq-engine/demos $OQ_PREFIX/share/openquake/engine
 
 tar -C ${OQ_ROOT}/${OQ_REL} -cpzvf openquake-${OQ_ENGINE_DEV}.tar.gz openquake
-sed -i 's/%_SOURCE_%/'openquake-${OQ_ENGINE_DEV}.tar.gz'/g' install.sh
+
+OQ_ARCHIVE="s/%_SOURCE_%/openquake-${OQ_ENGINE_DEV}.tar.gz/g" 
+if [ "$BUILD_OS" == "macosx" ]; then
+    sed -i '' $OQ_ARCHIVE install.sh
+else
+    sed -i $OQ_ARCHIVE install.sh
+fi
 GZIP=-1 tar -cpzvf openquake-opt-${OQ_ENGINE_DEV}.tar.gz openquake-${OQ_ENGINE_DEV}.tar.gz install.sh
 
 exit 0
